#61 rename common utils to async
