+++ src/utils/index.ts
@@ -1,1 +1,1 @@
-export * from './common';
+export * from './async';
